Added create user and login view.
Currently able to create new users from new view but still receiving a 404 for logging in. This may be android specific due to cordova handling of http
requests, and may just require tweaking of white listing - still investigating.

Backend service is added. The new view uses this to login and create new users. This service is also used for all other requests to the API backend.